django/requirements.txt
gunicorn
environs
pandas
Django
whitenoise
dash_bootstrap_components
django-bootstrap4
plotly
dpd-static-support
django_plotly_dash
dash_renderer
channels
channels-redis
daphne
redis
django-redis
mysqlclient
social-auth-app-django
python-dotenv
